chore: continues to move list to server
