+++ packages/ui/src/elements/ListDrawer/DrawerContent.tsx
@@ -17,7 +17,7 @@ import { useConfig } from '../../providers/Config'
 import { DocumentInfoProvider } from '../../providers/DocumentInfo'
 import { usePreferences } from '../../providers/Preferences'
 import { RenderCustomComponent } from '../RenderCustomComponent'
-import DefaultList from '../../views/List/Default'
+import { DefaultList } from '../../views/List'
 import formatFields from '../../views/List/formatFields'
 import { useDocumentDrawer } from '../DocumentDrawer'
 import Pill from '../Pill'
