+++ packages/ui/src/exports/views.ts
@@ -1,2 +1,2 @@
 export { DefaultDashboard } from '../views/Dashboard'
-export { DefaultList } from '../views/List/Default'
+export { DefaultList } from '../views/List'
